:recycle: refactor v2 internals into field submodule
