+++ mindee/parsing/v2/field/dynamic_field.py
@@ -34,13 +34,13 @@ def get_field_type(raw_response: StringDict, indent_level: int = 0) -> DynamicFi
     """Get appropriate field types."""
     if isinstance(raw_response, dict):
         if "value" in raw_response:
-            field_file = import_module("mindee.parsing.v2.simple_field")
+            field_file = import_module("mindee.parsing.v2.field.simple_field")
             field_class = getattr(field_file, FieldType.SIMPLE.value)
         elif "items" in raw_response:
-            field_file = import_module("mindee.parsing.v2.list_field")
+            field_file = import_module("mindee.parsing.v2.field.list_field")
             field_class = getattr(field_file, FieldType.LIST.value)
         elif "fields" in raw_response:
-            field_file = import_module("mindee.parsing.v2.object_field")
+            field_file = import_module("mindee.parsing.v2.field.object_field")
             field_class = getattr(field_file, FieldType.OBJECT.value)
         else:
             raise MindeeApiV2Error(f"Unrecognized field format in {raw_response}.")
