+++ mindee/parsing/v2/field/inference_result_fields.py
@@ -1,7 +1,11 @@
 from typing import Dict
 
 from mindee.parsing.common.string_dict import StringDict
-from mindee.parsing.v2.dynamic_field import DynamicField, FieldType, get_field_type
+from mindee.parsing.v2.field.dynamic_field import (
+    DynamicField,
+    FieldType,
+    get_field_type,
+)
 
 
 class InferenceResultFields(Dict[str, DynamicField]):
